[nemo-qml-plugin-calendar] Allow searching for imported events. Contributes to JB#38487
This commit adds a (threaded worker) method which allows clients to
search for an event to determine if it has been imported/synced.
This is useful e.g. to find invitation events which have been synced
and may have a different local-database UID than the original UID
from the invitation email, where the original UID might be stored
as a custom property.

It is exposed via an InvitationQuery type.

Contributes to JB#38487