[nemo-qml-plugin-calendar] Quantize alarm time for display. Contributes to JB#33108
Previously the code required the alarm time to be specific values
or it would be displayed as ReminderNone (i.e. no reminder) in the
calendar UI.

This commit ensures that we quantize the alarm time to one of the
specific values for display, as arbitrary times can be set if the
event comes from e.g. sync rather than created on the device.