[nemo-qml-plugin-calendar] Use ResponseType if available to show calendar attendance. Contributes to JB#46554
The ResponseType returned by an EAS server indicates how the user
responded to a event invitation request (Accepted, Declined, Tentative).
The ownerStatus is set to this value if available, which is used by the
UI to display the user's attendance status.

The previous behaviour used the list of invited users and attendance
status assigned to each to determine the user's response. The EAS server
won't always provide these values. Moreover, if the user has multiple
email addresses for an account the address used for the invitation may
not match the account username, preventing the value from being picked
up. In case the value is available and set, it's still used in
preference to the ResponseType.