[nemo-qml-plugin-calendar] Always return stored CalendarEvent object even if not part of any notebook anymore.
The CalendarEvent objects created by the manager are stored in
the mEventObjects private array. This array is always growing.
When an event is not listed anymore in the mEvents array
because the notebook it belongs to is not visible anymore,
fetching its CalendarEvent object should still be valid.
So move the array search before checking that the event is
currently part of any notebook.