[nemo-qml-plugin-calendar] Use mKCal::Notebook color as canonical color. Contributes to JB#51280
Previously, the canonical color used for display purposes in the
jolla-calendar application for a particular notebook was stored
in a QSettings .ini file.

That allowed the user to define a color for a calendar locally, without
perturbing the color specified on the server.  However, it also meant
that the semantics of color selection were not easily discoverable,
and further, that synchronisation of the color selection would never
be performed.

This commit ensures that the synced color will be used as the color
displayed in the application, and that the color selection will be
stored into the notebook in the database.

Finally, it ensures that if no color is set, a default color is chosen
and stored into the notebook, to avoid the possibility that the color
of the primary notebook can change between reboots.