[nemo-qml-plugin-calendar] Ensure that deleted events are in memory
When running several delete actions in a row from the calendarapi object,
the storage is reloaded between each delete call. This results in
the memory calendar to be emptied after each delete action. When the
next delete action is arriving, the event to delete is not in memory
anymore and the deletion is aborted.