[nemo-qml-plugin-calendar] Store in an occurrence if its corresponding event is all day. Contributes to JB#55785
In the worker thread, the local allDay hashtable may not be complete.
It only reflects the all day status for the newly added events.
But eventOccurrences() is dealing with all events in memory,
including some previously loaded ones. The solution is to add a flag
to EventOccurrence structure storing if the occurrence is coming
from an all day event.