[nemo-qml-plugin-calendar] Save the event when accepting the invitation. Contributes to JB#55926
When we call updateAttendee() - we call the SqliteStorage::calendarIncidenceChanged()
and add the uid of the event to mIncidencesToUpdate. Then, when the CalendarAgendaModel
is updating, it calls SqliteStorage::Private::addIncidence() which ignores our event
because its uid contains in mIncidencesToUpdate, and event is not displayed. To remove
the uid of the event from mIncidencesToUpdate we need to call save() method.
But if ServiceHandler::sendResponse() returned false, we need to restore the attendees
status of the event.