[nemo-qml-pugin-calendar] Correct the day for an event. JB#56431
The device can be in any time zone.
Events must be arranged by day in order to be displayed correctly.
But the border of days in different zones is different,
so the event must fall on the day of the border of which in the time zone of the device.