[nemo-qml-plugin-calendar] Create exception on object creation.
Move the call to dissociateSingleOccurrence() to the
creation of the CalendarEventModification object. It
is respecting more the idea of KCalendarCore to
create the exception, modify it and finally add it to
the calendar. The addition is done then via the ::save()
method of the CalendarEventModification class.

This also allow to get the recurrenceId for the new
exception synchronously and make the use of
CalendarChangeInformation not necessary anymore.