[tests] Improve waiting. JB#57616
Use QTRY_COMPARE instead of comparing in a loop waiting a little bit at
a time and then doing QCOMPARE. This is able to tell if there is need to
wait just a little bit longer for the test to pass. Also since the
earlier timeout was three seconds its now five.

That last loop was strange, it was basically a three second delay. I
made it to use QTRY_VERIFY instead. I also changed the previous wait to
use similar non-racy method of waiting.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>