Avoid deprecated QDateTime constructor
The QDateTime(const QDate&) constructor is deprecated, so this uses
QDateTime(QDate.startOfDay()) or QDateTime(QDate.endOfDay()) as appropriate.
This also simplified one place in calendarworker.cpp in
CalendarWorker::eventOccurrences().  In that, an elaborate dance to get the
last second of the second date was replaced by simply using QDate.endOfDay().

In all cases this is done with a QT_VERSION_CHECK(5, 14, 0) so compiling
with earlier versions of QT will still work.

Signed-off-by: Ed Beroset <beroset@ieee.org>