[nemo-qml-plugin-calendar] Add a shift on recurrenceId in this and future cases
In KCalendarCore::OccurrenceIterator, the exceptions are
find matching the recurrenceId with the generated
occurence start time from timesInInterval() called on
the parent, whatever the value of the current this
and future exception.

Moreover, ExtendedCalendar::dissociateSingleOccurrence()
is checking that recurrenceId is a valid recurring time
date for the parent and not the current this and future
exception. Which is consistent with KCalendarCore so far.

Thus, when calling the dissociation on an occurrence
coming from a this and future exception, apply a shift
due to the exception when creating the recurrenceId.