[nemo-qml-plugin-calendar] Add a search model. Contributes to JB#60453
Add a search function to CalendarManager. It is based
on ExtendedStorage::search() method. The worker is
loading all matching incidences and retrieving the
corresponding instance identifiers. The later are
passed back to the manager via a new searchResults
signal. While the former are sent back viw the
existing dataLoaded signal.

Also add a model listing occurrence of a search
query. This model is based on a EventListModel
where the identifiers are given as the result of
a search call.