[nemo-qml-plugin-calendar] Make Event::instanceId sole identifier for events. Contributes to JB#60800
Remove the recurrenceId from the API, so instanceId
is able to uniquely identify an event, or an exception.
This is done internally by affecting the result of
instanceIdentifier() to Event::instanceId.

This would allow later to add a notebook UID
information to this instanceId and thus being
able to add multi-notebook support without
changing once again the API.