[nemo-qml-plugin-calendar] Clone existing parent when saving an exception.
When moving the dissociation code out of the
save function, it had the drawback that only
properties copied in the CalendarData::Event
structure are preserved in the exception.

This may creates some issues with the sync
plugins that may use properties that are not
part of CalendarData::Event structure.

Ensure that such properties are kept when
creating an exception by cloning the parent
before updating it with the exception
modifications.