[nemo-calendar] Avoid -WError build option. JB#63621
Kf5-calendarcore has deprecated notebook api which we still use,
removed in the Qt6 version. That causes build here to fail if
warnings are not tolerated. Even regardless of that the Werror
maybe better not used since it will lead to problems like this
for other changes outside the module.